Feat: Initial solutions to day 1.
Day 1 puzzles were solved in a REPL and the solutions are simply copied
from the IPython history. I might tidy them up later, perhaps depending
on where the rest of the advent goes. I could add tests/CLI interface.
I'm also not 100% sold on the structure. But this is fine for now.