Feat: Add solutions for day 12 parts 1 & 2
Much better with bread first search over depth first search. I had to
look up the implementation, but happy with how it turned out. In part 2
it's also possible to exit the BFS after finding the first candidate
start point, as it's guaranteed to be a start point with the shortest
path to the end. Still very messy.